MIM-133 拆分新会话表单职责
